Add ncbiseg output to xml first working version
Officially, this now seems to be a valid scheme, however we might have
to add some more EDAM references, and maybe we could include more
information.

Scores are currently not commented in, because they would require a
'value' field, and there is nothing senseful we could put there.

Something is also strange about the namespaces. bx: works now, but that
is only for bioxsd, our own, doesn't work yet